updating Position information response

+++ openair2/COMMON/f1ap_messages_types.h
@@ -712,7 +712,7 @@ typedef struct f1ap_resource_set_type_semi_persistent_s {
 
 typedef struct f1ap_resource_set_type_aperiodic_s {
   uint8_t sRSResourceTrigger; //(M)
-  uint8_t slotoffset; //(M)
+  long slotoffset; //(M)
 } f1ap_resource_set_type_aperiodic_t;
 
 typedef union f1ap_resource_set_type_c {
@@ -780,7 +780,7 @@ typedef struct f1ap_srs_resource_s {
   uint8_t c_SRS; //(M)
   uint8_t b_SRS; //(M)
   uint8_t b_hop; //(M)
-  uint8_t groupOrSequenceHopping; //(M) neither	= 0, groupHopping	= 1, sequenceHopping	= 2
+  long groupOrSequenceHopping; //(M) neither	= 0, groupHopping	= 1, sequenceHopping	= 2
   f1ap_resource_type_t resourceType; //(M) choice
   uint16_t slotOffset; //(M)
   uint16_t sequenceId; //(M)
